Change version requirement for mediafile from ~=0.6.0 to >=0.6.0
The currently required version 0.6.0 of the mediafile package is really
old (current version is 0.12.0). In addition, the mediafile features
used by extrafiles are minimal. So it should be safe to require any
version greater than 0.6.0 (in particular since the version spec for
beets itself is also just >=1.4.7).

This change avoids an explicit downgrade of the mediafile package
version installed by beets itself.